Give all cards their own Edit/View components
