+++ src/components/profile/edit/EditTags.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    <header class="profile__section-header">
+      <h2>Tags</h2>
+    </header>
+    <p>This section cannot yet be edited in DinoPark</p>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'EditTags',
+};
+</script>
